==> notebooks/02_data_cleaning.ipynb <==
## Data Cleaning

import pandas as pd

df = pd.read_csv('../data/original/data.csv')

irrelevant_columns = [
    'Order',
    'PID',
    'Alley',
    'Low Qual Fin SF',
    'Enclosed Porch',
    '3Ssn Porch',
    'Pool Area',
    'Pool QC',
    'Misc Feature',
    'Misc Val',
    'Mo Sold',
    'Utilities',
    'Street',
    'Heating',
    'Land Slope',
    'Land Contour',
    'MS SubClass'
]

df.drop(columns=irrelevant_columns, inplace=True)

temp = (df.isna().sum()/len(df)*100).sort_values(ascending=False)
nan_cols = temp[temp>40].index.tolist()
nan_cols

df.drop(columns=nan_cols, inplace=True)

df.columns = df.columns.str.strip().str.replace(' ', '_').str.lower()

df.duplicated().sum()

df.info()

### Refactor the type's of the columns

df[['1st_flr_sf', '2nd_flr_sf', 'gr_liv_area','wood_deck_sf', 'open_porch_sf', 'saleprice', 'lot_area']] = df[['1st_flr_sf', '2nd_flr_sf', 'gr_liv_area','wood_deck_sf', 'open_porch_sf', 'saleprice', 'lot_area']].astype('float')

df[['bsmt_full_bath', 'bsmt_half_bath', 'garage_cars', 'garage_yr_blt']]= df[['bsmt_full_bath', 'bsmt_half_bath', 'garage_cars', 'garage_yr_blt']].astype('Int64')

df.to_parquet('../data/processed/02_data_cleaned.parquet', index=False)
