==> notebooks/03_missing_values_imputation.ipynb <==
## Missing Values Imputation

from importlib import reload
import pandas as pd
import imputation as imp
import numpy as np
reload(imp)

df = pd.read_parquet('../data/processed/02_data_cleaned.parquet')
df = df.replace({None: np.nan})

### Analysis

imp.stats_summary(df, df.select_dtypes(include=['float']).columns, soft=False)

imp.stats_summary(df, df.select_dtypes(include=['int']).columns, soft=False)

numeric_columns = df.select_dtypes(include=['float']).columns.union(pd.Index(['overall_qual', 'overall_cond', 'year_built', 'year_remod/add', 'totrms_abvgrd', 'screen_porch']))
int_columns = ['bsmt_full_bath', 'bsmt_half_bath', 'full_bath', 'half_bath', 'bedroom_abvgr', 'kitchen_abvgr', 'fireplaces', 'garage_cars', 'yr_sold', 'garage_yr_blt']
categorical_columns = df.select_dtypes(include=['object']).columns

### FLoat columns Imputation

imp.stats_summary(df, numeric_columns, soft=True)

transformation_recommendation = imp.transformation_recommender(df, numeric_columns, soft=False, display_info=False)
transformation_recommendation

reload(imp)
recommendations = imp.imputer_recommender_float(
    df,
    transformation_recommendation,
    soft=False,
    display_info=False
)

recommendations

reload(imp)
df_imputed = imp.apply_imputation_strategy(df, recommendations, display_plot=False)

### Int and Categorical Columns with a High Number of Unique Values

df_imputed = imp.impute_with_catboost(df_imputed, int_columns, categorical_columns)

df_imputed.to_parquet('../data/processed/03_data_imputed.parquet', index=False)
